Move Microsoft Graph credential logic to ee/
Premium features belong in ee/, so the Graph credential service implementation
moves to ee/server/service/microsoft_graph_credentials.go. server/service keeps
the endpoint plumbing and free-tier stubs, matching the certificate authorities
split.

msGraphClientFactory moves from the core Service to the ee Service, since the
premium implementation is now the only consumer and an unexported field is not
reachable through the embedded fleet.Service.

windowsEntraGUIDRegex is hoisted from server/service/appconfig.go to
fleet.IsValidEntraGUID. Both the core app config validation and the premium
credential validation need it, and duplicating it would let the two drift.

The free-tier ApplyMicrosoftGraphCredentials stub returns nil for an empty list
rather than a license error. GitOps reaches this endpoint on every run, sending
an empty list whenever default.yml omits microsoft_graph_credentials, so a
free-tier run that never configured Microsoft Graph must succeed. This mirrors
BatchApplyCertificateAuthorities.

No behavior change.
