+++ ee/server/service/mdm_external_test.go
@@ -100,7 +100,6 @@ func setupMockDatastorePremiumService(t testing.TB) (*mock.Store, *eeservice.Ser
 		nil,
 		nil,
 		nil,
-		nil,
 	)
 	if err != nil {
 		panic(err)
@@ -130,6 +129,7 @@ func setupMockDatastorePremiumService(t testing.TB) (*mock.Store, *eeservice.Ser
 		nil,
 		nil,
 		nil,
+		nil, // msGraphClientFactory
 	)
 	if err != nil {
 		panic(err)
