+++ server/service/microsoft_graph_credentials_test.go
@@ -370,24 +370,6 @@ func TestListMicrosoftGraphCredentials(t *testing.T) {
 		"the read must not decrypt secrets it is about to mask")
 }
 
-func TestMicrosoftGraphVerifyMessage(t *testing.T) {
-	t.Parallel()
-	for _, tc := range []struct {
-		name     string
-		err      error
-		contains string
-	}{
-		{"permission", &msgraph.Error{StatusCode: http.StatusForbidden}, "DeviceManagementServiceConfig.Read.All"},
-		{"auth", &msgraph.Error{StatusCode: http.StatusUnauthorized}, "rejected the credential"},
-		{"transient", &msgraph.Error{StatusCode: http.StatusBadGateway}, "temporarily unavailable"},
-		{"non-graph error", errors.New("dial tcp: timeout"), "Couldn't connect to Microsoft Graph"},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			assert.Contains(t, microsoftGraphVerifyMessage(tc.err), tc.contains)
-		})
-	}
-}
-
 func TestMicrosoftGraphCredentialsAuth(t *testing.T) {
 	t.Parallel()
 	env := setupGraphCredsTest(t, fleet.TierPremium, "test-private-key", nil)
